Fix review comment: move all new staff from utils_ranges_sycl.h into parallel_backend_sycl_reduce_then_scan_pos_tools.h
